// src/api/generated/models/CreateViewEventRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Запрос на создание события просмотра
 */
export type CreateViewEventRequest = {
  entity_type: string;
  entity_id: number;
  listing_pos?: number;
  listing_page?: number;
  contragent_phone?: string;
};