// src/api/generated/models/LocationsListResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MarketplaceLocation } from './MarketplaceLocation';

export type LocationsListResponse = {
  locations: Array<MarketplaceLocation>;
  count: number;
  page: number;
  size: number;
};