// src/api/generated/models/MarketplaceAddToCartRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MarketplaceOrderGoodCustom } from './MarketplaceOrderGoodCustom';

export type MarketplaceAddToCartRequest = {
  contragent_phone: string;
  good: MarketplaceOrderGoodCustom;
};