// src/api/generated/models/MarketplaceCartResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { MarketplaceOrderGood } from './MarketplaceOrderGood';

export type MarketplaceCartResponse = {
  contragent_phone: string;
  goods: Array<MarketplaceOrderGood>;
  total_count: number;
};